Fix test-suite-source-ordering-field error
Summary: Fix test-suite-source-ordering-field error. Diff to land the changes onto master branch.

Test Plan: Run complex-dag-cow.yaml twice, once with timestamp as source-ordering-field and once with test-suite-source-ordering-field. While running with test-suite-source-ordering-field use TestSuiteFileBasedSchemaProvider as schemaprovider-class

Reviewers: balajee, pwason, PHID-PROJ-pxfpotkfgkanblb3detq

Reviewed By: balajee, pwason, PHID-PROJ-pxfpotkfgkanblb3detq

JIRA Issues: HUDI-755

Differential Revision: https://code.uberinternal.com/D6010175

+++ hudi-integ-test/src/main/java/org/apache/hudi/integ/testsuite/schema/TestSuiteFileBasedSchemaProvider.java
@@ -56,7 +56,7 @@ public class TestSuiteFileBasedSchemaProvider extends FilebasedSchemaProvider {
       fields.add(newField);
     }
     Schema.Field sourceOrderingField =
-        new Schema.Field(SchemaUtils.SOURCE_ORDERING_FIELD, Schema.create(Type.INT), "", 0);
+        new Schema.Field(SchemaUtils.SOURCE_ORDERING_FIELD, Schema.create(Type.LONG), "", 0);
     fields.add(sourceOrderingField);
     Schema mergedSchema = Schema.createRecord(schema.getName(), schema.getDoc(), schema.getNamespace(), false);
     mergedSchema.setFields(fields);
